Narrow the build script's module graph
`build.rs` recompiles part of the library so it can call
`cli::Cli::command()` for man-page generation. It did so by declaring
`src/cli/mod.rs`, which dragged in the entire `cli` subtree — merging,
discovery, diagnostics, localized value parsing — plus `cli_l10n`,
`host_pattern`, `output_mode`, and `theme`. Almost none of that is
reachable from the build script, so five module-wide
`#[expect(dead_code, ...)]` attributes were needed to silence 110
unused-item diagnostics.

Those suppressions were not merely noisy. They also masked genuinely
dead code: an unused `pub` item in `src/cli/config.rs` was reported by
neither compilation unit, because the library exports that module
publicly and the build script's expectation covered the rest.

Replace the five declarations with a facade: an inline `cli` module in
`build.rs` naming exactly the three files the Clap schema needs. To make
that slice self-contained, split the library along the same seam:

- `src/cli/command.rs` holds the Clap definitions, previously the top
  half of `src/cli/parser.rs`. `src/cli/parser.rs` keeps the
  localization-aware parsing entry point.
- `src/cli/preferences.rs` holds the four `Cli` accessors that resolve
  output policy, which were the only reason the build script compiled
  `theme` and, transitively, `output_mode`.
- `src/cli/validation.rs` holds `MAX_JOBS` and `validation_error`, so
  `src/cli/config.rs` no longer reaches up into `src/cli/mod.rs`.
- `src/host_matching.rs` holds `HostCandidate` and `HostPattern::matches`,
  the only items in `src/host_pattern.rs` the schema does not need.

All five expectations are removed and the build script compiles with no
unused-item diagnostics. The generated man page is byte-identical, the
public API is unchanged, and no dependency was added. Rerun directives now
track the modules actually compiled. Adding a dependency outside the slice
surfaces as a build-script compile error, which is the intended signal;
`docs/developers-guide.md` records the boundary.

Closes #513
