+++ src/cli/discovery.rs
@@ -12,7 +12,7 @@ use std::path::{Path, PathBuf};
 use std::sync::Arc;
 use tracing::{debug, debug_span};
 
-use super::parser::Cli;
+use super::command::Cli;
 
 #[path = "discovery_diagnostics.rs"]
 mod diagnostics;
