+++ src/host_pattern.rs
@@ -1,7 +1,10 @@
 //! Shared host pattern validation helpers.
 //!
-//! The module centralises normalisation and matching logic so CLI parsing and
-//! runtime policy evaluation agree on allowable host syntax.
+//! The module centralises host pattern normalisation so CLI parsing and
+//! runtime policy evaluation agree on allowable host syntax. Matching a
+//! concrete hostname against a parsed pattern lives in
+//! `crate::host_matching`, which keeps this module's dependency surface
+//! narrow enough for `build.rs` to compile it for man-page generation.
 
 use crate::localization::{self, LocalizedMessage, keys};
 use serde::{Deserialize, Serialize};
@@ -12,15 +15,6 @@ use thiserror::Error;
 struct HostPatternInput<'a>(&'a str);
 
 impl<'a> HostPatternInput<'a> {
-    const fn as_str(self) -> &'a str {
-        self.0
-    }
-}
-
-#[derive(Copy, Clone)]
-pub(crate) struct HostCandidate<'a>(pub(crate) &'a str);
-
-impl<'a> HostCandidate<'a> {
     const fn as_str(self) -> &'a str {
         self.0
     }
@@ -218,20 +212,6 @@ impl HostPattern {
             wildcard,
         })
     }
-
-    pub(crate) fn matches(&self, candidate: HostCandidate<'_>) -> bool {
-        let host = candidate.as_str().to_ascii_lowercase();
-        if self.wildcard {
-            // Wildcard patterns match only subdomains, not the apex domain.
-            // Example: "*.example.com" matches "sub.example.com" but not
-            // "example.com".
-            host.strip_suffix(&self.pattern)
-                .and_then(|prefix| prefix.strip_suffix('.'))
-                .is_some_and(|prefix| !prefix.is_empty())
-        } else {
-            host == self.pattern
-        }
-    }
 }
 
 impl<'a> TryFrom<&'a str> for HostPattern {
@@ -284,7 +264,7 @@ impl<'de> Deserialize<'de> for HostPattern {
 
 #[cfg(test)]
 mod tests {
-    //! Unit tests for host pattern parsing and wildcard matching.
+    //! Unit tests for host pattern parsing and normalisation.
     use super::*;
 
     use anyhow::{Result, ensure};
@@ -302,26 +282,6 @@ mod tests {
         ensure!(
             parsed.wildcard == wildcard,
             "expected wildcard {wildcard} for pattern {pattern}",
-        );
-        Ok(())
-    }
-
-    #[rstest]
-    #[case("example.com", "example.com", true)]
-    #[case("example.com", "sub.example.com", false)]
-    #[case("*.example.com", "sub.example.com", true)]
-    #[case("*.example.com", "example.com", false)]
-    #[case("*.example.com", "deep.sub.example.com", true)]
-    #[case("*.example.com", "other.com", false)]
-    fn host_pattern_matches_expected(
-        #[case] pattern: &str,
-        #[case] host: &str,
-        #[case] expected: bool,
-    ) -> Result<()> {
-        let parsed = HostPattern::parse(pattern)?;
-        ensure!(
-            parsed.matches(HostCandidate(host)) == expected,
-            "expected match={expected} for {host} against {pattern}",
         );
         Ok(())
     }
